# mountinfo: simpify, speedup, and fix unescape