@kolyshkin 2024-07-26
Basically, the differences between this and #143 are:
 - #143 is way simpler and thus easier to maintain down the road;
 - this one is slightly more future-proof (in case kernel or some weird fs decides to escape some more characters).