@kolyshkin 2024-08-14
Rebased; check that golangci-linter v1.60.1 (from #147) produces no warnings.