@kolyshkin 2024-09-06
PTAL @thaJeztah, I think I've addressed all your comments.